+++ src/Dto/Webhook/Webhook/Create.php
@@ -26,7 +26,7 @@ use Shopgate\ConnectSdk\Dto\Webhook\Webhook;
 
 /**
  * @method Create setName(string $name)
- * @method Create setEndpoint(string $endpoint)
+ * @method Create setEndpoint(string $endpoint) - full endpoint URI to call once an event gets triggered
  * @method Create setActive(bool $active)
  * @method Create setEvents(Dto\Event[] $events)
  *